+++ src/extensions/rules/rulesext.cpp
@@ -118,7 +118,7 @@ RulesClassExtension::RulesClassExtension(const RulesClass* this_ptr) :
     ComesNearWaypointDistance(CELL_LEPTON_W * 5),
     IsAIDetectDisguise(true),
     IsAIOneHarvesterInSingleplayer(true),
-    IsCloakedTechnosTriggerCellTags(false)
+    IsCellTagsIgnoreStealth(true)
 {
     //if (this_ptr) EXT_DEBUG_TRACE("RulesClassExtension::RulesClassExtension - 0x%08X\n", (uintptr_t)(ThisPtr));
 
@@ -293,7 +293,7 @@ void RulesClassExtension::Object_CRC(CRCEngine &crc) const
     crc(IsAIDetectDisguise);
     crc(AIHarvestersPerRefinery.Count());
     crc(IsAIOneHarvesterInSingleplayer);
-    crc(IsCloakedTechnosTriggerCellTags);
+    crc(IsCellTagsIgnoreStealth);
 }
 
 
@@ -726,7 +726,7 @@ bool RulesClassExtension::General(CCINIClass &ini)
     MaxBeacons = ini.Get_Int(GENERAL, "MaxBeacons", MaxBeacons);
     SelfHealingCap = ini.Get_Float(GENERAL, "SelfHealingCap", SelfHealingCap);    
     SelfHealingRate = ini.Get_Float(GENERAL, "SelfHealingRate", SelfHealingRate);
-    IsCloakedTechnosTriggerCellTags = ini.Get_Bool(GENERAL, "CloakedTechnosTriggerCellTags", IsCloakedTechnosTriggerCellTags);
+    IsCellTagsIgnoreStealth = ini.Get_Bool(GENERAL, "CellTagsIgnoreStealth", IsCellTagsIgnoreStealth);
 
     /**
      *  Allow replacing any signle movement zone with a copy of RA2's water MZone.