+++ src/extensions/rules/rulesext.h
@@ -277,5 +277,5 @@ public:
     /**
      *  Determines whether the cloaked technos can trigger cell tags when used with Entered By trigger event.
      */
-    bool IsCloakedTechnosTriggerCellTags;
+    bool IsCellTagsIgnoreStealth;
 };